% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCutoffVars.R
\name{addCutoffVars}
\alias{addCutoffVars}
\title{Add cutoff variables}
\usage{
addCutoffVars(obj, vars, fun = identity, type = "mG", plotOnly = F, ...)
}
\description{
Cutoff right of the global maximum
}
