% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateByWell.R
\name{aggregateByWell}
\alias{aggregateByWell}
\title{Aggregate data per well}
\usage{
aggregateByWell(data, fun = median, onlyCellNumber = F)
}
\arguments{
\item{data}{}
}
\value{
dataframe containing aggregated data per well
}
\description{
Aggregate data per well
}
