% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisCellCycleFractions.R
\name{analysisCCFractions}
\alias{analysisCCFractions}
\title{Analysis of cell cycle fractions}
\usage{
analysisCCFractions(data, aggregate = F)
}
\arguments{
\item{data}{data}
}
\description{
desc
}
