% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Analyze data}
\usage{
analyze(
  data,
  frm = as.formula(VAL ~ GRP + DOSE + (1 | VERSUCH/PLATTE/TREATMENT)),
  frm0 = as.formula(VAL ~ 1 + (1 | VERSUCH/PLATTE/TREATMENT)),
  ref = "Ko_1",
  nCores = NULL,
  trSep = "_"
)
}
\arguments{
\item{data}{data.frame as obtained by the aggregateByWell() 
or zTransform() functions}

\item{frm}{Formula to use with lme4::lmer}

\item{frm0}{Null-model formula}

\item{ref}{Reference level of the GRP variable}
}
\description{
TAVI data analysis, consinstig of a linear mixed
effect modelling approach with cross validation.
}
