% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeNLME.R
\name{analyzeNLME}
\alias{analyzeNLME}
\title{Analyze data}
\usage{
analyzeNLME(
  data,
  rand = as.formula(1 | VERSUCH/PLATTE/TREATMENT),
  frm = as.formula(VAL ~ GRP + DOSE),
  frm0 = as.formula(VAL ~ 1),
  ref = "Ko_1",
  nCores = NULL,
  trSep = "_",
  level = NULL
)
}
\arguments{
\item{data}{data.frame as obtained by the aggregateByWell() 
or zTransform() functions}

\item{frm}{Formula to use with lme4::lmer}

\item{frm0}{Null-model formula}

\item{ref}{Reference level of the GRP variable}

\item{level}{level used in predict() fit an nlme object}
}
\description{
TAVI data analysis, consinstig of a linear mixed
effect modelling approach with cross validation.
}
