% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeTAVI.R
\name{analyzeTAVI}
\alias{analyzeTAVI}
\title{Analyze TAVI data}
\usage{
analyzeTAVI(
  data,
  frm = as.formula(VAL ~ GRP + DOSE + (1 | VERSUCH/SUBJ/TREATMENT)),
  frm0 = as.formula(VAL ~ 1 + (1 | VERSUCH/SUBJ/TREATMENT)),
  ref = "preTAVI"
)
}
\arguments{
\item{data}{data.frame as obtained by the aggregateByWell() 
or zTransform() functions}

\item{frm}{Formula to use with lme4::lmer}

\item{frm0}{Null-model formula}

\item{ref}{Reference level of the GRP variable}
}
\description{
TAVI data analysis, consinstig of a linear mixed
effect modelling approach with cross validation.
}
