% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignCutoff.R
\name{assignCutoff}
\alias{assignCutoff}
\title{Assigns Cutoff result to each cell}
\usage{
assignCutoff(
  data,
  calc,
  type = "CIexcl",
  var = "Intensity_IntegratedIntensity_DNA.nucl",
  nCores = NULL,
  outvar = NULL,
  ...
)
}
\arguments{
\item{data}{data}

\item{calc}{Data calculated with cellCycleFractIntegrDNAInt()}

\item{type}{Method on how to assign cell cycle status
to each cell. CIexcl assigns only cells with values 
outside the confidence interval for the separating value 
determined in cellCycleFractIntegrDNAInt() by bootstrapping.
CImedian uses the median as cutoff.}
}
\description{
A
}
