% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignNcFilterRes.R
\name{assignNcFilterRes}
\alias{assignNcFilterRes}
\title{Assign ncAreaFilter Results}
\usage{
assignNcFilterRes(
  data,
  calc,
  type = "CImedian",
  varC = "AreaShape_Area.cell",
  varN = "AreaShape_Area.nucl",
  nCores = NULL,
  ...
)
}
\description{
des
}
