% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCutoffs.R
\name{calcCutoffs}
\alias{calcCutoffs}
\title{Calculate cutoffs for filters}
\usage{
calcCutoffs(obj, fun = c("cc", "nc", "fb"), delim = "\\t", rmTreatCol = c(1))
}
\description{
Calculate cutoffs for filters
}
