% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisCellCycleFractions.R
\name{calcFract}
\alias{calcFract}
\title{Calculate fractions (cellcycle)}
\usage{
calcFract(data, cuts, var = "Intensity_IntegratedIntensity_DNA.nucl", ...)
}
\description{
Calculate fractions (cellcycle)
}
