% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSyntheticVars.R
\name{calculateSyntheticVars}
\alias{calculateSyntheticVars}
\title{Calculate Synthetic Variables}
\usage{
calculateSyntheticVars(data, mapping, fun = "mean")
}
\arguments{
\item{P}{}
}
\description{
Calculate Synthetic Variables
}
