% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellCycleFractIntegrDNAInt.R
\name{cellCycleFractIntegrDNAInt}
\alias{cellCycleFractIntegrDNAInt}
\title{Derive cell cycle fractions from nuclear 
DNA intensity measurements}
\usage{
cellCycleFractIntegrDNAInt(
  data,
  var = "Intensity_IntegratedIntensity_DNA.nucl",
  nBoot = 100,
  lowerCL = 0.1,
  upperCL = 0.9,
  log = F,
  cutUpper = NULL,
  xMinGlobMax = 0,
  xMaxGlobMax = NULL,
  no_cores = NULL,
  ..
)
}
\arguments{
\item{data}{data.frame containing at least the columns 
TREATMENT, VERSUCH, PLATTE and the variable used for 
DNA cell cycle inference (see parameter var)}

\item{var}{Variable in which the nuclear integrated dna 
intensity is stored. Default is 
"Intensity_IntegratedIntensity_DNA.nucl"}

\item{nBoot}{number of runs to determine minima/maxima
to derive cutoffs}

\item{lowerCL}{lower confidence level, default: 0.1}

\item{upperCL}{upper confidence level, default: 0.9}

\item{cutUpper}{use only data below a certain cutoff, 
default: NULL -> not used}
}
\value{
data.frame containing the computed values
}
\description{
Calculates cell-cycle fractions (G1, G2) 
of cells based on a nuclear DNA intensity variable; 
defaults to "Intensity_IntegratedIntensity_DNA".
}
