% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgExperimentClass.R
\name{checkFilter}
\alias{checkFilter}
\title{Check and filter treatments}
\usage{
checkFilter(obj, filter = NULL)
}
\arguments{
\item{filter}{remove treated categories with small number of cells,
defaults to NULL (no filtering)}
}
\description{
Check and filter treatments
}
