% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corBetweenRep.R
\name{corBetweenRep}
\alias{corBetweenRep}
\title{correlation between replicates}
\usage{
corBetweenRep(data, filter = c("Meta", "Number", "VERSUCH", "TREAT"))
}
\arguments{
\item{data}{specifies the input data}

\item{filter}{drops columns}
}
\description{
Only numeric columns will be used;
Irrespective of substance concentration
}
