% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAdditionalVars.R
\name{createAdditionalVars}
\alias{createAdditionalVars}
\title{creation of new variables (ratios or differences)}
\usage{
createAdditionalVars(data, type = "both", vars, order = c("nucl", "cell"))
}
\arguments{
\item{data}{a}

\item{type}{one of "both", "diff", "ratio"}

\item{vars}{a}

\item{order}{which value to use first (nucl/cell; nucl-cell) -> order=c("nucl", "cell")}
}
\description{
creates variables calculated as ratios or differences of two other variables
}
