% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarPlots.R
\name{drawRadarplots}
\alias{drawRadarplots}
\title{Draw Radarplots}
\usage{
drawRadarplots(
  data,
  vars,
  agg = "median",
  res = 250,
  file = "",
  width = 1000,
  height = 1000,
  labels = F,
  col = NULL,
  ctrlLevel = "Ko_1",
  main = NULL,
  pPos = 1.05,
  pSym = "*",
  pTest = "t.test",
  pAdj = "none",
  pCol = list(`0.05` = "red", `0.1` = "blue"),
  pCex = 3,
  na.rm = T,
  scale = T
)
}
\arguments{
\item{data}{specifies the data for plotting}

\item{vars}{specifies which variables are used for plotting and the order of plotting}

\item{agg}{statistical method of data aggregation (median or mean)}

\item{main}{Title}
}
\value{
no return value
}
\description{
Draw Radarplots
}
