% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgExperimentClass.R
\name{getData}
\alias{getData}
\title{Load experiment data from}
\usage{
getData(obj, force = F, ...)
}
\arguments{
\item{obj}{imgExp instance}
}
\description{
reads cell profiler output and treatment files
}
