% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSingleCol.R
\name{getNRows}
\alias{getNRows}
\title{Get number of rows from file}
\usage{
getNRows(file)
}
\arguments{
\item{file}{Filename}
}
\description{
Get number of rows from file

Number of rows (including header)
}
