% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearDecomposition.R
\name{getResiduals}
\alias{getResiduals}
\title{Get Residuals}
\usage{
getResiduals(data, vars, selVar)
}
\arguments{
\item{data}{d}

\item{selVar}{already selected variables}

\item{var}{names of the vars to use}
}
\value{
a
}
\description{
D
}
