% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMissingCutoffs.R
\name{imputeMC}
\alias{imputeMC}
\title{Impute missing cutoffs}
\usage{
imputeMC(cuts, plot = F, ...)
}
\arguments{
\item{cuts}{output of the cellCycleFractIntegrDNAInt function}

\item{plot}{plot estimates}
}
\description{
Fills missing data
}
