% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgExperimentClass.R
\name{initialize,imageExp-method}
\alias{initialize,imageExp-method}
\title{Constructor imageExp}
\usage{
\S4method{initialize}{imageExp}(.Object, expID = character, experiment = list)
}
\arguments{
\item{.Object}{imageExp object}

\item{expID}{experiment ID}

\item{experiment}{two level list containing experimental
data folders}
}
\value{
imageExp instance
}
\description{
Constructor for imageExp
}
