% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectLmeCor.R
\name{intersectLmeCor}
\alias{intersectLmeCor}
\title{A Intersect lme/cor results}
\usage{
intersectLmeCor(corR, lmeR, corCut = NULL, lmePCut = NULL, nRes = NULL)
}
\arguments{
\item{corR}{Ergebnisse der corBetweenRep() Funktion}

\item{lmeR}{Ergebnisse der lmeKoVsTreatment() Funktion}

\item{corCut}{Cutoff for a minimum required correlation (pearson)}

\item{p-Value}{Cutoff for lme derived pval (maximum)}
}
\description{
A
}
