% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearDecomposition.R
\name{linearDecomp}
\alias{linearDecomp}
\title{Linear Decomposition}
\usage{
linearDecomp(
  data,
  versuche,
  vars = NULL,
  startvar = "AreaShape_Area.x",
  maxIterations = 40,
  continueIfNA = TRUE
)
}
\arguments{
\item{data}{data}

\item{versuche}{Zu vergleichende Versuche}

\item{vars}{output of intersectLmeCor}

\item{startvar}{Startvariable, standard: AreaShape_Area.x}

\item{maxIterations}{maximum of allowed Iterations}

\item{continueIfNA}{If the residual cannot be calculated, NA values occur.
The calculation continues if set to TRUE. If set to FALSE it stops.}
}
\value{
data.frame containing bla ...
}
\description{
This function correlates the residuals of
}
