% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmeKoVsTreatment.R
\name{lmeKoVsTreatment}
\alias{lmeKoVsTreatment}
\title{LME Kontrolle vs Treatment}
\usage{
lmeKoVsTreatment(
  data,
  vars = "all",
  adjust = "none",
  koVar = "Ko_1",
  koGrep = "Ko"
)
}
\arguments{
\item{data}{}

\item{vars}{can be all or a list of selected colnames}

\item{adjust}{which posthoc adjustment should be used?
Can be "tukey", standard is "none"}

\item{koVar}{a}

\item{koGrep}{b}
}
\description{
Use VERSUCH as random factor
}
