% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncAreaFilter.R
\name{ncAreaFilter}
\alias{ncAreaFilter}
\title{Nucl/Cell Ratio Filter}
\usage{
ncAreaFilter(
  data,
  varC = "AreaShape_Area.cell",
  varN = "AreaShape_Area.nucl",
  lowerCL = 0.1,
  upperCL = 0.9,
  nBoot = 100,
  xCut = 15,
  bw = 0.1,
  xMinGlobMax = 1.5,
  ...
)
}
\arguments{
\item{data}{data}
}
\description{
desca
}
