% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelationMatrix.R
\name{plotCorrelationMatrix}
\alias{plotCorrelationMatrix}
\title{Plot correlation matrix}
\usage{
plotCorrelationMatrix(
  data,
  vars = "all",
  selVars = NULL,
  filter = c("Meta", "Object", "Parent", "PLATTE", "VERSUCH"),
  perTreatment = F
)
}
\arguments{
\item{data}{d}

\item{vars}{v}

\item{filter}{f}
}
\description{
Plot correlation matrix
}
