% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVerlauf.R
\name{plotVerlauf}
\alias{plotVerlauf}
\title{Plotte Verlaufskurve}
\usage{
plotVerlauf(dat, ylim = NULL, column = NULL, main = "", fun = median, lwd = 1)
}
\arguments{
\item{column}{zu plottende Spalte der Daten. 
Wenn null, Ausgabe der Zellzahl (=Anzahl Zeilen / Well)}

\item{fun}{Funktion zur Aggregation der Daten, 
Standard: Median}
}
\value{
A
}
\description{
Vlerauf
}
