% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc}
\alias{qc}
\title{Create QC plots}
\usage{
qc(
  data,
  folder = NULL,
  file = NULL,
  uid = NULL,
  vars = "nCells",
  fun = NULL,
  na.rm = T
)
}
\arguments{
\item{data}{data.frame containing the cell profiler data to check. 
Expects the columns TREATMENT, PLATTE, VERSUCH, Metadata_Well}

\item{file}{pdf filenamename to store the analysis at}

\item{vars}{Variables to analyze, defaults to nCells (number of cells)}

\item{fun}{Function to obtain aggregated values per well, e.g. median}

\item{na.rm}{delete na in fun}
}
\description{
Creates QC plots
}
