% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{Read data}
\usage{
readData(
  data,
  vars = NULL,
  treatVar = c("Treatment", "Konzentration"),
  wellVar = "well",
  files = c("Treatment.csv", "Cells.txt", "Primarieswithoutborder.txt",
    "Cytoplasm.txt"),
  normalizeWells = T,
  sepTreat = "_",
  adaptVars = F,
  force = F,
  filterVars = c("Intensity_IntegratedIntensity_DNA", "Intensity_MedianIntensity_DNA",
    "AreaShape_Area")
)
}
\arguments{
\item{data}{list contianing a lists of filesnames, all corresponding 
a separate experiment}

\item{vars}{Variables to read, defaults to NULL (all variables)}

\item{treatVar}{Treatment / Grouping variables as given in the 
treatment file}

\item{files}{CellProfiler output files and treatment file}

\item{normalizeWells}{convert A01->A1}

\item{force}{force continuation if dimensions don't match}

\item{well}{Well Variable within the treatment file}
}
\description{
Reads experiment associated data. Variables get suffix 
.cell, .nucl, .cyto
}
