% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSingleCol.R
\name{readSingleCol}
\alias{readSingleCol}
\title{Read CP file single column}
\usage{
readSingleCol(file, var, nrow = NULL, type = "numeric", delim = "\\t")
}
\arguments{
\item{file}{CP file}

\item{var}{CP column name}
}
\description{
Read CP file single column
}
