% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCutoffs.R
\name{calcCutoffs}
\alias{calcCutoffs}
\title{Calculate cutoffs for filters}
\usage{
calcCutoffs(obj, cutoffPrev = NULL, delim = "\\t", rmTreatCol = c(1))
}
\arguments{
\item{cutoffPrev}{returen data from calcCutoffs from previous run}
}
\description{
Calculate cutoffs for filters
}
